Align file dimension types
Agent-Logs-Url: https://github.com/bernardosantiago44/MapcelRepositorioArticulos/sessions/aaff8dd9-f0e3-4278-882f-f1d199af9a3d

+++ MapcelRepositorioArticulos/Models/File.cs
@@ -14,8 +14,8 @@ public class FileAsset
     public Guid CompanyCode { get; set; }
     public IReadOnlyList<string> LinkedArticles { get; set; }
     public Uri? ThumbnailUrl { get; set; }
-    public long? Width { get; set; }
-    public long? Height { get; set; }
+    public int? Width { get; set; }
+    public int? Height { get; set; }
 
     // Empty Constructor (for serialization or EF Core)
     public FileAsset()
@@ -40,8 +40,8 @@ public class FileAsset
         Guid companyCode, 
         IReadOnlyList<string> linkedArticles,
         Uri? thumbnailUrl,
-        long? width,
-        long? height)
+        int? width,
+        int? height)
     {
         Id = id;
         Name = name;
